Merge the develop branch to the master branch, preparation to v2.3.3

+++ test/testContracts/ERC677ReceiverTest.sol
@@ -1,15 +1,14 @@
 pragma solidity ^0.4.19;
 
-import "../../contracts/ERC677Receiver.sol";
-
+import "../../contracts/interfaces/ERC677Receiver.sol";
 
 contract ERC677ReceiverTest is ERC677Receiver {
     address public from;
-    uint public value;
+    uint256 public value;
     bytes public data;
-    uint public someVar = 0;
+    uint256 public someVar = 0;
 
-    function onTokenTransfer(address _from, uint256 _value, bytes _data) external returns(bool) {
+    function onTokenTransfer(address _from, uint256 _value, bytes _data) external returns (bool) {
         from = _from;
         value = _value;
         data = _data;
@@ -17,7 +16,7 @@ contract ERC677ReceiverTest is ERC677Receiver {
         return true;
     }
 
-    function doSomething(uint _value) public {
+    function doSomething(uint256 _value) public {
         someVar = _value;
     }
 }
